requirements.txt
boto3
botocore
Cython==0.29.37
einops
einops-exts==0.0.4
Flask==2.2.5
Flask-RESTful==0.3.10
ftfy==6.1.1
h5py==3.7.0
hydra-core
imageio==2.23.0
jieba==0.42.1
kiwisolver==1.4.5
kornia==0.7.3
kornia_rs==0.1.5
llvmlite==0.43.0
lmdb==1.5.1
matplotlib==3.7.2
msgpack==1.0.8
networkx==3.2.1
ninja==1.11.1.1
nltk==3.8.1
numba==0.60.0
numpy==1.25.2
omegaconf
open_clip_torch
# open3d
opencv-python
opencv-python-headless
pandas==2.2.2
pathtools==0.1.2
Pillow==9.3.0
plyfile==0.7.4
protobuf==3.20.1
pycocotools==2.0.6
pynvml==11.4.1
pypng==0.20220715.0
pyviz3d==0.2.32
PyYAML==5.4.1
scikit-fmm==2024.5.29
scikit-learn==1.2.0
scipy==1.9.3
sentence-transformers==3.0.1
sentencepiece==0.2.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorboardX==2.5.1
timm==1.0.7
tokenizers==0.19.1
torchdiffeq==0.2.4
torchmetrics==0.11.0
torchsde==0.2.6
tqdm==4.64.1
transformers==4.44.0
transforms3d==0.4.1
triton==2.1.0
urllib3==2.2.2
yacs==0.1.8
yapf==0.32.0
supervision
